Avoid locking analyzer assemblies when loading them.
We expect that many users will check analyzer assemblies into source control; we've done this ourselves in the Roslyn solution. This poses a problem: any source control operations that would modify the assembly on disk will fail while VS or VBCSCompiler are running.

To avoid this problem we will handle loading assemblies ourselves, by first copying them file into a byte array and then loading from there. This avoids the need to shadow-copy files (and thus the need to clean them up) which is our other option to avoid the locking.

The interesting bit is that we need to handle loading the assemblies dependencies as well, by subscribing to the AppDomain.AssemblyResolve event. (changeset 1296284)