Fix issue where the completion list wasn't displaying the correct list of symbols after a comma in the generic argument list of a reduced extension method.
Class1<string, int>.Create().Test<string,

In this case, the user's partially typed code will parse as a less-than-expression and the comma is parsed as skipped-token-trivia attached to a missing semicolon. It turns out that our current syntactic context checks are permissive enough to allow this case. However, when we check to the SemanticModel.LookupSymbols() to see if there are any symbols with arity that match the identifier to the left (i.e. "Test" in the example above), we didn't specify the "includeReducedExtensionMethods" parameter. The fix is to do exactly that. (changeset 1302205)