Rewrite of the native client to improve stability and give better error messages.
The new native client is written in C++11 to be more maintainable and forward-looking. The native client is a static library, which is now linked into the csc2 and vbc2 projects instead of having a single csc2.cpp file which is shared between the two projects and built for each based on a preprocessor definition.

The MSBuildTask is also configured to use csc2/vbc2 and feature no extra code for Roslyn, which will prevent us from having to maintain both a native and a managed version of the client. (changeset 1307711)