Fixing 530858: Generate property: leads to CS0053 Inconsistent Accessibility
When a property is generated inside an internal type but it's return type is protected, we were generating a public property which leads to inconsistent accessibility. While computing minimum accessibility, we were treating internal as more restrictive than protected which is not how the compiler sees it. I've changed it to this order of most restrictive to least restricitve : private > protectedandinternal > protected > internal > protectedorinternal > public. I've tests the matrix of combinartions and they all seem to produce compilable code now.  (changeset 1308601)