Propagate project renames to the workspace.
When a project is renamed, we need to fetch updated display names and file paths for it. We leave the project ID as is, with the debug path which is still the old string, since we have no good way to change that.

If this change works, DustinCa should get the credit since he wrote most of this. (changeset 1309226)