Disallow string literals in #pragma warnings directives and allow identifiers instead. Also stop validating warning ids supplied in #pragma warning directives. These were both decisions that were made at a design meeting a few weeks back. The corresponding VB changes (for #Disable) have already been implemented - this change implementss the same for C#.
Note: Parsing / lexing of identifiers inside #pragma warning directives is identical to that inside #define directives except for the following difference. Inside #define, very long identifiers are truncated to 128 characters to maintain backwards compatibility with previous versions of the compiler (Dev12-). Since support for identifiers inside #pragma warning directivess is new, we don't have any backwards compatibility constraints for this. So we preserve the identifier exactly as it appears in source. (changeset 1309687)