Some cleanup of IL temps created when emitting addresses. Results of EmitAddress should not be ignored.
Generally C# allows only taking address of expressions that are "addressable". In rare cases an expression is not directly addressable and a temp must be created. However generally we know exactly the lifetime of such reference and the corresponding temp.
Not knowing the life-time and leaking a local intentionally is extremely rare and only happens in specific kinds of synthesized code, so we know it is safe thing to do.

In all other cases the temp is either unexpected (assert was added) or its lifetime is known (release code was added).
 (changeset 1311067)