Fix for bug 798375: Rename : Microsoft.CodeAnalysis.VisualBasic.Simplification.VisualBasicSimplificationService.Expander.VisitSimpleName
While trying to rename a parameter of a lambda field initializer, the speculative model created for references to the parameter within the lambda body was being passed in a position within the lambda. As chaining of speculative models is not supported (see won't fixed bug 797122), we cannot create a speculative model for individual statements within a lambda initializer. We need to speculate on the entire field initializer. The rename rewriter was correctly walking up the nodes to find the topmost node that can be speculated, i.e. the field initializer syntax; however it wasn't adjusting the position for speculation causing the new field initializer to be speculated in context of the lambda body rather then the entire field initializer. (changeset 1313031)