Fix for bug 995855: [CodePlex] C# - Roslyn suggests dubious refactoring options for integral type casts.
Fix one of the error cases for this bug: We incorrectly report unncessary cast within compound assignment expressions. The language specs requires that an implicit conversion exists from the right side of the compound assignment expression to the left type. We didn't perform this check in the speculation analyzer, causing us to break cast simplification for these cases. (changeset 1314225)