Fix for bug 995855 (fixes #245): C# - Roslyn suggests dubious refactoring options for integral type casts.
Fix for second incorrect cast removal suggestion reported in this bug: If the conversion from expression type to cast type and also from cast type to outer type are both explicit numeric, then one of these conversions is a narrowing conversion causing loss of precision. Removal of cast might lead to code that compiles fine if the conversion from expression type to outer type is implicit numeric. However, the code leads to a different output as we no longer have a narrowing conversion. Hence, we should not report this scenario as an unncessary cast. (changeset 1314339)