Delete CodeHierarchy.
This was a prototype feature that is mostly (but not entirely!) disabled. Since there is no plans to ship it, fully remove it so it's not a code maintenance burden going forward, and doesn't surprise us by the parts that were still being loaded through MEF. If we do intend to ship it, we can rescue it from source control.
***NO_CI***
 (changeset 1314571)