Bug 1000457: SuppressSyntaxDiagnosticsOnEventAddAccessorBasic - compiler unit test failing on build servers
This seems to be an intermittent test failure while executing the diagnostic analyzers, but the exception is swallowed by the analyzer driver. We cannot do much to fix the failure unless we have a dump, so I have added debug only code to rethrow the swallowed exception, so we can diagnose the failure when it occurs again. (changeset 1314572)