Bugfix 987373
Introduce an option, ShouldUseFormattingSpanCollapse,  to tell the formatting engine if it should collapse the formatting spans if there are more than 30 spans to be formatted(Collapsing involes taking the a bunch of spans and returning one span whose startingPoint is startingPoint of the first span and the ending point is the endpoint of the last span in the bunch). The default value is to not collapse

There are few CodeFix Provider which act on the whole document for which the formatting engine will collapse the spans to one span.

In order  to improve telemetry, I have introduced some new CodeActions instead of using the generic Document/SolutionChangeCodeAction (changeset 1317739)