Add more diagnostics for intermittently failing unit test (Bug 1000457: SuppressSyntaxDiagnosticsOnEventAddAccessorBasic - compiler unit test failing on build servers)
This unit test seems to be failing intermittently on build servers for release builds. I added some debug only code (changeset 1314572) for better diagnostics of this failure, but that turns out to be useless due to release only failures. So I am rolling back changeset 1314572 and adding unit test framework support for configuring if the analyzer driver should handle exceptions or not. The default behavior for unit tests would be to *not* catch any driver/analyzer exceptions. However, tests can explicitly flip his behavior for testing exception scenarios.
This should enable us to get a stack dump when this test fails again on build servers. (changeset 1317998)