Update VBCSCompiler.exe to watch for analyzer changes.
If an analyzer assembly changes on disk after we've loaded it, we need to stop accepting connections and allow VBCSCompiler.exe to shut down. Since we've already loaded the assembly we can't load it again* and future builds will then run with an out-of-date analyzer.

When we see that an analyzer has changed on disk we simply close the pending PipeStream and do not recreate it. However, the closing of the pending PipeStream is what signals VBCSCompiler.exe to shutdown, so we also need to make sure we wait for all in-progress connections to finish up before we exit. In this way we both prevent new compilation requests from using the old analyzer, and exit the VBCSCompiler process as quickly as possible.

*We could load it again if the assembly identity changes, e.g. if the version number increments. However, that is an optimization we can make later if need be.  (changeset 1318480)