Bugfix 866094 - "SymbolId doesn't support VB Module when it is created from csharp side"
VB Modules are interpreted in C# as symbols with TypeKind Class (instead of TypeKind Module on the VB side), so NamedTypeSymbolKey.Resolve did not realize that these should match up.  There's no compelling reason to check the TypeKind in NamedTypeSymbolKey, so we can remove this check entirely rather than add a special case for this scenario. (changeset 1321956)