Bugfix 995021 Proper indentation with Complex trivia
When calculating the indentation for the line with leading complex trivia(not in the same line) the calculated indentation has to be used instead of the existing spaces.

In VB, there are cases where the complex trivia (colon trivia) is still part of the same line which makes the calculation indentation useless. (changeset 1324609)