Port C# synthesized symbol refactoring work to VB
This is a part of porting C# synthesized refactoring work where the names of long-lived variables follow a naming pattern that allows us to reverse-engineer the SynthesizedLocalKind value from the variable name. Especially in this change, we give a synthesized variable an explicit name based on its SynthesizedLocalKind at emit time, not at local symbol creation time. (changeset 1324862)