Fix for bug 961128: Pending work items for lightbulb support for suppressions
This change fixes the global/assembly suppress message attribute generation for member symbols via the light bulb. This attribute takes a Target string denoting a string representation for the symbol on which the suppression is to be applied. Currently, we just walk up to the enclosing type and emit its fully qualified name, suppressing the diagnostic on entire type rather then the member.
Now, we do the correct thing by using a unique string for member. This unique string for declaration symbols is generated using the existing DocumentationCommendId.CreateDeclarationId API, which is used across workspaces by various features that need such a string representation. This type also exposes GetSymbolsForDeclarationId API for converting the string representation back to symbols. The suppression logic in compiler layer uses this API to decode this attribute's target string. In the process I also fixed couple of bugs in the DocumentationCommendId APIs which surfaced through the added tests. (changeset 1325198)