Fix 529531: Conflict resolution engine in refactorings should insert parentheses to keep correct parsing
In this bug, the expander is trying to expand code like "x as N + 1" where N is an alias to int?. The expander was expanding this to x as int? + 1 which is not the fully expanded form. Fixing it to expand to x as global::System.Nullable<int> + 1 which when simplified knows to put parentheses correctly: (x as int?) + 1 (changeset 1328490)