DevDiv #1019283: Missing local variables while debugging
The PDB has an attributes field for each local.  We were setting "1" for "compiler-generated" and "0" for everything else.  It turns out that dev12 was expecting "1" for "ignored by the expression compiler" and "0" for everything else.  This was most important for display class locals, which are compiler-generated but should definitely not be ignored by the expression compiler.

As far as I can determine, we now emit the same values as the native compiler.

TODO: The Roslyn expression compiler does not consume the attributes field - it just uses the name to determine whether a local is interesting.  Should we consume the field for consistency with the native expression compiler?

CR: PHarring, KevinH, NGafter, VSadov (changeset 1329523)