Do not format within the braces if they're on the same line for array/collection/object initializer expressions. This is a heuristic to prevent brace completion from breaking user expectation/muscle memory in common scenarios.
In Dev12 , with auto brace completion enabled, typing { in array, collection, object initializers does not format the brace session range (from { to }). this means users are used to typing in a space right after '{'. In Roslyn, we used to format this and put in the space, thus breaking muscle memory in  common typing scenarios. With this change, we revert to Dev12 behavior.

some examples:
No formatting cases:
1. new int[] {}
2. var x = new { Age = 1 }

Formatting still happens for :
class C {}
int Prop {}

The changes themselves are inside
1. BraceCompletion where we format the tracking span
2. SmartTokenFormatting which gets triggered for TypeChar of '{'

The change in both places is the same - range to format does not include the '}' (closebrace) token. (changeset 1331104)