Last /warnaserror flag on the command line should always win.
Note: C# and VB behaviors are different for /warnaserror. VB always allows /warnaserror-:xyz to override /warnaserror (i.e. specific option always overrides general one). C# on the other hand allows the last of these options specified on the command line to win - so if /warnaserror:xyz is followed by /warnaserror, then xyz is promoted to an error - but if /warnaserror:xyz follows /warnaserror, then xyz stays as a warning. We can't make the behavior for the two languages identical because we need to maintain backwards compatibility. (changeset 1331612)