Using the metadata reader to parse analyzer type names before loading them.
The AnalyzerFileReference type is responsible for loading a dll and instantiating individual analyzer types. We were enumerating all available types through reflection. This was the root cause of very poor error messages for diagnostic authors while writing analyzers. Since we are loading an assembly to even determine what language the analyzers are for, we end up loading more assemblies than are needed in the VS process. This change solves both of those issues by making the following changes:

- Read the file using the metadata reader to determine all types that have the DiagnosticAnalyzerAttribute. Build a map of language -> analyzer typenames so that we can avoid loading the assembly if there are no types for the language of interest.
- The DiagnosticAnalyzerAttribute used to take a params array of strings - changing it to be a single string with allow multiple to keep the parsing simple.
- Extract the error handling code out to an event so that hosts can report errors encountered during analyzer loading.

 (changeset 1331893)