DevDiv #1019283: Part 2 - Consuming the attributes field on local debug info
On its own, this change seems kind of silly - instead of consuming a mangled name from the PDB, we set the name to null.  The main benefit is that other PDB producers have a way to communicate to the EE that certain locals should not be consumed, regardless of how they are named. (changeset 1332108)