DevDiv #1015887: Marshalling System.DateTime for ISymUnmanagedWriter2.DefineConstant2
Formerly, we were just using .NET's built-in object-to-variant marshalling.  Unfortunately, .NET marshals DateTimes as the number of days since 1899/12/30, whereas the native VB compiler marshalled them as the number of ticks since the Unix epoch (i.e. a much, much larger number).  This resulted in incorrect display of DateTime constants in the EE.

The solution is to do our own marshalling between object and variant.  We use Marshal.GetNativeVariantForObject for every type but DateTime.

Thanks to JasonMal for helping me debug the COM marshalling!

CR: ChuckS, GreggM, KevinH, VSadov (changeset 1334950)