PERF: VBSyntaxKind allocations
Reduce boxing of VB SyntaxKind instances. This shows up in a few places, primarily when IndexOf or Contains (which called IndexOf) is called on an array of SyntaxKinds. The implementation of IndexOf uses the default equality comparer and, since SyntaxKind is an enum, the default equality comparer is an "ObjectEqualityComparer".
This is 1.7% of allocations in the VB typing test. (changeset 1337531)