Show method body diagnostics for partial method implementations.
Consider:

File A.cs:

  partial class MyPartialClass
  {
      partial void MyPartialMethod(MyUndefinedType m);
  }

File B.cs:

  partial class MyPartialClass
  {
      partial void MyPartialMethod(MyUndefinedType m)
      {
          c = new MyUndefinedType(23, true);
      }
  }

Currently, we show squiggles (and corresponding items in the Error List) under "MyUndefinedType" in the signatures of both the partial method declaration (in A.cs) and implementation (in B.cs). However, we show no squiggles for any errors in the body of the implementation. We expect two: one under "c" as it is an undefined variable, and another under "MyUndefinedType" in the object creation.

The issue is in how we decide which symbols to analyze when collecting method body diagnostics. When analyzing B.cs, we pass in a filter that causes us to ignore everything defined outside of that file. The symbol we get by default for MyPartialMethod, however, is only considered to be defined in A.cs. As such we filter it out instead of checking for diagnostics.

When we come across a symbol representing a source method the filter needs to explicitly check if it is partial. If it is, then it must be the symbol for the declaration, and we need to ask it for the symbol for the implementation and check that against the filter as well. (changeset 1338052)