Fix 1029549: Clean up TriggerDiagnostic Category
We have a TriggerDiagnosticDescriptor.TriggerCategory which was overloaded to mean two things - a trigger diagnostic and the ruleset editor doesn't show these diagnostics. Some analyzers like RudeEditAnalyzer and AddImports were using this category to mean that they shouldn't be configurable in the ruleset editor. This change adds a new tag called NotConfigurable which would tell the ruleset editor to not display the rule in the ruleset editor. TriggerDiagnosticDescriptors by definiton wouldn't get displayed either.
 (changeset 1338925)