Show analyzer ids for VSIX analyzers in ruleset files
For project analyzers we return a dictionary<string, ienumerable<analyzers>> to the ruleset editor so that it can show the analyzer name corresponding to the individual rules. We don't do that for the workspace analyzers. This change fixes that.

Also, I'm making the categories localizable and changing them to be simply "Design" instead of "Microsoft.Design" per the other code review i sent out. (changeset 1342176)