Bugfix 1036954 : Compiler only expects IOException and BadImageFormatException coming out of PortableExecutableReference.GetMetadata.
Any other exception (like UnauthorizedAccessException) needs to be wrapped into IOException, unless it already derives from IOException.
     VisualStudioMetadataReferenceManager.GetStorageInfoFromTemporaryStorage directly calls File.Open which may throw all sorts of exceptions.

The fix is to call FileUtilities.OpenStream which wraps the exceptions properly

Since this is behavior expected any where in the product, the APIs Open, Create have now been wrapped with the helpers to generate the IOException instead of some random exception. (changeset 1347659)